Add files via upload
This is one of my projects for a clothing store sales including different states, gender, age over a timespan of almost 3-5 years. You'll be able to see the sorted data with advanced and interactive pivot table analysis to make these data more visually appealing and easy to understand. The graphs and charts can be customized according to the interest of the clients and so on.